+++ dataland-frontend/tests/e2e/specs/search-company-framework-data/SearchCompaniesForFrameworkDataDropdownFilter.ts
@@ -22,6 +22,8 @@ before(function () {
 });
 
 describe("As a user, I expect the search functionality on the /companies page to adjust to the selected dropdown filters", () => {
+  const failureMessageOnAvailableDatasetsPage = "We're sorry, but your search did not return any results.";
+
   it(
     "The framework filter should contain SFDR even though it is not yet implemented, and synchronise " +
       "between the search bar and the URL",
@@ -82,7 +84,7 @@ describe("As a user, I expect the search functionality on the /companies page to
       )
         .wait("@companies-meta-information")
         .get("div[class='col-12 text-left']")
-        .should("contain.text", "We're sorry, but your search did not return any results.")
+        .should("contain.text", failureMessageOnAvailableDatasetsPage)
         .get("#country-filter")
         .click()
         .get('input[placeholder="Search countries"]')
@@ -111,7 +113,7 @@ describe("As a user, I expect the search functionality on the /companies page to
       cy.visit(`/companies?input=${demoCompanyToTestFor.companyName}&sector=${demoCompanyWithDifferentSector.sector}`)
         .wait("@companies-meta-information")
         .get("div[class='col-12 text-left']")
-        .should("contain.text", "We're sorry, but your search did not return any results.")
+        .should("contain.text", failureMessageOnAvailableDatasetsPage)
         .get("#sector-filter")
         .click()
         .get('input[placeholder="Search sectors"]')
@@ -211,7 +213,7 @@ describe("As a user, I expect the search functionality on the /companies page to
           });
           cy.visit(`/companies?input=${companyName}`)
             .get("div[class='col-12 text-left']")
-            .should("contain.text", "We're sorry, but your search did not return any results.");
+            .should("contain.text", failureMessageOnAvailableDatasetsPage);
         }
       );
 
@@ -246,7 +248,7 @@ describe("As a user, I expect the search functionality on the /companies page to
             .should("exist");
           cy.visit(`/companies?input=${companyName}&framework=${DataTypeEnum.EutaxonomyNonFinancials}`)
             .get("div[class='col-12 text-left']")
-            .should("contain.text", "We're sorry, but your search did not return any results.");
+            .should("contain.text", failureMessageOnAvailableDatasetsPage);
           cy.visit(
             `/companies?input=${companyName}&framework=${DataTypeEnum.EutaxonomyNonFinancials}&framework=${DataTypeEnum.EutaxonomyFinancials}`
           )
